refactor(headless-styles): move deepMerge to its own file
