+++ packages/headless-styles/sandbox_old/plugins/css-to-ts/convert-rules/composes.js
@@ -1,3 +1,5 @@
+import deepMerge from '../utils/deepMerge'
+
 export default function composes(styleObj) {
   for (const [className, value] of Object.entries(styleObj)) {
     if (typeof value !== 'object') {
@@ -10,46 +12,22 @@ export default function composes(styleObj) {
 
     if (composes) {
       if (composes in styleObj) {
-        deepMerge(styleObj[className], styleObj[composes])
+        styleObj[className] = deepMerge(styleObj[className], styleObj[composes])
       } else if (composes.includes(' from ')) {
         // Preserve value as array for later
         value.composes = [composes]
-        deepMerge(styleObj[className], value)
+        styleObj[className] = deepMerge(styleObj[className], value)
       } else {
         composes.split(/\s+/).forEach((otherKey) => {
-          deepMerge(styleObj[className], styleObj[otherKey])
+          styleObj[className] = deepMerge(
+            styleObj[className],
+            styleObj[otherKey]
+          )
         })
       }
     }
-    deepMerge(styleObj[className], overrides)
+    styleObj[className] = deepMerge(styleObj[className], overrides)
   }
 
   return styleObj
-}
-
-function deepMerge(source, target) {
-  for (const [key, value] of Object.entries(target)) {
-    // Overwrite primitive values, merge objects together
-    if (key in source && isObject(value)) {
-      deepMerge(source[key], value)
-    } else if (isObject(value)) {
-      source[key] = { ...value }
-    } else if (Array.isArray(value)) {
-      source[key] = mergeArrays(source[key], value)
-    } else {
-      source[key] = value
-    }
-  }
-}
-
-function isObject(item) {
-  return item && typeof item === 'object' && !Array.isArray(item)
-}
-
-function mergeArrays(source, target) {
-  if (Array.isArray(source)) {
-    return [...source, ...target]
-  }
-
-  return target
 }
